Use an abstract compiler pass instead
