+++ core-bundle/src/DependencyInjection/Compiler/AbstractConfigureFilesystemPass.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of Contao.
+ *
+ * (c) Leo Feyer
+ *
+ * @license LGPL-3.0-or-later
+ */
+
+namespace Contao\CoreBundle\DependencyInjection\Compiler;
+
+use Contao\CoreBundle\DependencyInjection\Filesystem\ConfigureFilesystemInterface;
+use Contao\CoreBundle\DependencyInjection\Filesystem\FilesystemConfiguration;
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+
+abstract class AbstractConfigureFilesystemPass implements CompilerPassInterface, ConfigureFilesystemInterface
+{
+    public function process(ContainerBuilder $container): void
+    {
+        $this->configureFilesystem(new FilesystemConfiguration($container));
+    }
+}
