--- core-bundle/src/DependencyInjection/Filesystem/AbstractFilesystemBundle.php
@@ -1,32 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This file is part of Contao.
- *
- * (c) Leo Feyer
- *
- * @license LGPL-3.0-or-later
- */
-
-namespace Contao\CoreBundle\DependencyInjection\Filesystem;
-
-use Symfony\Component\DependencyInjection\Container;
-use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
-use Symfony\Component\HttpKernel\Bundle\AbstractBundle;
-
-/**
- * @experimental
- */
-abstract class AbstractFilesystemBundle extends AbstractBundle implements ConfigureFilesystemInterface
-{
-    public function getContainerExtension(): ExtensionInterface|null
-    {
-        if ('' === $this->extensionAlias) {
-            $this->extensionAlias = Container::underscore(preg_replace('/Bundle$/', '', $this->getName()));
-        }
-
-        return $this->extension ??= new FilesystemBundleExtension($this, $this->extensionAlias);
-    }
-}
