+++ phpstan.neon
@@ -52,7 +52,6 @@ parameters:
         - identifier: staticMethod.internal
         - identifier: staticMethod.internalClass
         - identifier: classConstant.internalClass
-        - identifier: class.extendsInternalClass
 
         # Ignore @no-named-arguments errors caused by PHPUnit.
         # https://github.com/sebastianbergmann/phpunit/issues/6163
